Updated docs and fixed a typo in README
