+++ src/actionMetadataKey.ts
@@ -1,3 +1,4 @@
-/** @internal */
+/** @ignore */
 export const actionMetadataKey = Symbol('when-action');
+/** @ignore */
 export const inputMetadataKey = Symbol('when-input');
